+++ doc/Changelog.md
@@ -3,7 +3,9 @@
 # git master
 
 * [50](https://github.com/Eyescale/vmmlib/pull/50):
-  Improved matrix and lowpass filter API
+  Improved matrix, quaternion and lowpass filter API
+* [49](https://github.com/Eyescale/vmmlib/pull/49):
+  Add lookAt support for 4x4 matrices
 
 # Release 1.10 (07-04-2016)